+++ src/runtime/components/Sections/admin.vue
@@ -431,9 +431,10 @@
                             <div class="btn-text">{{ $t("createGlobal") }}</div>
                           </button>
                           <button
+                            v-if="!guideConfig.disabled"
                             ref="intro-find-more-blobal"
                             class="intro-find-more-blobal hp-button globalTour"
-                            @click="runIntro('globalTour', true)"
+                            @click="runIntro('globalTour', true, null, 'global-content-guide-btn')"
                           >
                             <div class="btn-text intro">?</div>
                           </button>
@@ -450,9 +451,10 @@
 
                   <div class="flexSections top-bar-page-content-wrapper">
                     <button
+                      v-if="!guideConfig.disabled"
                       ref="intro-relaunch"
                       class="intro-relaunch hp-button"
-                      @click="runIntro('topBar', true)"
+                      @click="runIntro('topBar', true, null, 'relaunch-guide-btn')"
                     >
                       <div class="btn-text intro">?</div>
                     </button>
@@ -1784,6 +1786,30 @@ const dynamicSideBarComponentPath = ref('')
 const sectionsOptionsComponentPath = useState('sectionsOptionsComponentPath', () =>(''))
 
 const currentThemeTab = ref({})
+
+const guideConfig = useState('guideConfig', () => {
+  const defaultConfig = {
+    disabled: false,
+    autoStart: true,
+    override: false
+  }
+  try {
+    const hooksJs = importJs('/js/global-hooks') // assuming this is sync
+    if (hooksJs?.guide_config) {
+      const config = hooksJs.guide_config()
+      if (config && typeof config === 'object') {
+        return {
+          ...defaultConfig,
+          ...config
+        }
+      } else return defaultConfig
+    } else {
+      return defaultConfig
+    }
+  } catch {
+    return defaultConfig
+  }
+})
 
 // Computed properties
 const activeVariation = computed(() => {
@@ -3464,7 +3490,7 @@ const initiateIntroJs = async () => {
           }
           currentPages.value = response.data.current_pages
           if (currentPages.value !== null && currentPages.value === 0) {
-            if (pageNotFound.value) {
+            if (pageNotFound.value && guideConfig.value.autoStart === true) {
               await runIntro('createPage') // Await intro run
             }
           }
@@ -3482,7 +3508,14 @@ const initiateIntroJs = async () => {
     }
   } catch {} // Outer catch remains for potential errors before API call
 }
-const runIntro = async (topic, rerun, lastSavedTopic) => {
+const runIntro = async (topic, rerun, lastSavedTopic, action) => {
+  if (guideConfig.value.disabled === true) {
+    return
+  }
+  if (guideConfig.value.override === true) {
+    await nuxtApp.callHook('start-guide', topic, action)
+    return
+  }
   if (intro.value && topic === 'globalTour') {
     intro.value.setDontShowAgain(true)
     useCookie('intro-last-step').value = null